requirements.txt
numpy # Для научных вычислений и работы с массивами
pandas # Для обработки и анализа данных, особенно CSV-файлов
scikit-learn # Для утилит машинного обучения (например, MinMaxScaler, f1_score)


matplotlib # Для создания статических графиков
tqdm # Для отображения прогресс-баров в консоли
torchinfo # Для вывода сводной информации о моделях PyTorch
plotly # Для создания интерактивных графиков в веб-интерфейсе


streamlit # Для создания интерактивных веб-приложений (pred_web.py, train_web.py)


pyodbc # Драйвер для подключения к MS SQL Server
psycopg2-binary # Драйвер для подключения к PostgreSQL


python-telegram-bot # Для отправки сообщений и фото через Telegram